requirements.txt
# TN InternFair - Python Dependencies
# Install: pip install -r requirements.txt

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.1

# Auth & Security
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4

# Validation
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Environment
python-dotenv==1.0.0

# AI/ML (Local AI evaluation using Ollama)
crewai
langchain-ollama
langchain
langchain-community
requests

# Mock Data Generation
faker==22.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
aiofiles==23.2.1
python-dateutil==2.8.2
